+++ backend/internal/handlers/helpers_test.go
@@ -132,7 +132,7 @@ func (ah testAppHandler) SetDewControllerEnabled(enabled bool) error {
 	return nil
 }
 
-func (ah testAppHandler) SetTargetTemperature(targetTemperature int) error {
+func (ah testAppHandler) SetTargetTemperature(targetTemperature float64) error {
 	return nil
 }
